Move to Stack.ts, add console default
