+++ plugins/scenario/worker/Config.ts
@@ -2,6 +2,7 @@
 export const scenarioGlob = 'scenario/**.ts';
 export const workerCount = 3;
 export const defaultFormats = {
+  console: {},
   json: {
     output: 'scenario-results.json'
   }
